// src/app/models/tenant.ts
export interface TenantModel {
  id?: any;
  title?: any;
  createAt?: any;
  address?: any;
  companyId?: any;
  imageUrl?: any;
  username?: any;
  password?: any;
  name?: any;
  description?: any;
  getData?: any;
  remoteCode?: any;
  isActive?: any;
  isBlocked?: any;
  isVerified?: any;
  isDeleted?: any;
  openDate?: any;
  closeDate?: any;
  vatType?: any;
  currency?: any;
  other?:any;
}